+++ packages/annotation-toolkit/release-notes.md
@@ -1,6 +1,8 @@
 ### vNext
 - Show helpful message when `<AppShell />` is missing a layers prop.
-- Fix bug where actions menu header was showing even when there were no gathered actions.
+- Fix bug where duplicate actions were showing for selected alwaysOn layers.
+- Add a '⬩' marker next to layers in the actions menu that are selected by virtue of selection, as opposed to the `alwaysOn` prop. Helpful for debugging and intuiting why certain actions are shown.
+- Add a `hideName` boolean prop to prevent layers from showing up in the Layers Panel. Default: `false`.
 
 ### v1.0.5
 - [See here for initial release notes](https://github.com/facebookresearch/Mephisto/pull/427)